Add files via upload
To Run this Program First You have to save the all file in same directory. And Open the app.py file just run it. After running the command or file Flask will start a development server on your local machine, and you'll see some output in the terminal indicating that the Flask application is running. You might see messages like "Running on http://127.0.0.1:5000" or "Debugger PIN: XXX-XXX-XXX", which indicate that the Flask server is running and listening for incoming requests.

Once the Flask application is running, you can open a web browser and visit the URL http://127.0.0.1:5000 (or whatever URL is shown in the terminal output) to access your Flask application. This will take you to the homepage of the website, as defined in the home.html template.
You can then navigate through the different pages of our website by clicking on Start Quiz and it will take you to the another page that is quiz page if you click on the Yes, Start Quiz it will directly takes you to the problems page and afterwards result page. 
To stop the Flask application, you can press Ctrl + C in the terminal where the Flask server is running. This will stop the server, and you'll no longer be able to access your Flask application until you run the app.py file again.